requirements.txt
Django==4.1.2
djangorestframework
django-filter
django-cors-headers
django-import-export
djangorestframework-simplejwt
dj-rest-auth
gunicorn
Pillow
pybit
channels
channels-redis
